Lee Elliott:
here's an update for the AN-225.

As well as re-tuning to fix the incidence problem I also re-wrote
most of the auto-pilot configuration, especially the altitude
hold modes where I've tried to re-use controllers.  This needs a
Nasal script to monitor the standard A/P locks and set the
appropriate locks for the new controller structure.