Joacim Persson:
Added steering to the four aftmost main gear axels -- with animations!

Found out just the other day that the AN-225 indeed steers on some of the
main gear wheels. The minimum turning radius is a wild guess however.
(Assumed a 45 deg average on the nose gears.) It now turns (quite
effortlessly) around a point straight out to the side from the second main
gear axel (nr 2 of the 3 non-steerable main gear axels), with a minimum
radius of 25.75 meters. (i.e. near engine nr 1 or 6) The option "squared"
was removed on steering. (Squaring radians messed up my calculations of the
turning angle of the respective wheels. Not that it's a perfect calculation
now either.) Ground wheel steering should, rather, not be connected to
directly the rudder pedals I think...

Also added transition-time(s) for the steering.